+++ nn.h
@@ -22,15 +22,15 @@ typedef struct nn_layer_t {
   size_t num_neurons_;        //num neurons in layer
   size_t weights_per_neuron_; //based on num neurons in prev layer connected
 
-  long double* biases_;            //neuron bias
-  long double* outputs_;           //output of neuron
-  long double* weighted_sums_;     //weighted sum of neuron z + bias
-  long double** weights_;          //2d array rows for each neuron
+  float* biases_;            //neuron bias
+  float* outputs_;           //output of neuron
+  float* weighted_sums_;     //weighted sum of neuron z + bias
+  float** weights_;          //2d array rows for each neuron
 
-  long double** avg_weight_grads_; //average gradients for every weight
+  float** avg_weight_grads_; //average gradients for every weight
 
-  long double* errors_;            //neuron error delta
-  long double* avg_errors_;        //stores avg error over a mini batch
+  float* errors_;            //neuron error delta
+  float* avg_errors_;        //stores avg error over a mini batch
 
 } nn_layer_t;
 
@@ -63,14 +63,14 @@ bool initNNet(neural_network_t* n_net, size_t num_layers,
  * If the pointer passed is null, the verification step is skipped.
  */
 bool sgdNNet(neural_network_t* n_net,
-    long double* const samples,
-    long double* const expected,
+    float* const samples,
+    float* const expected,
     size_t num_samples,
     uint64_t epochs,
-    long double eta ,
+    float eta ,
     size_t mini_batch_size,
-    long double* verif_samples,     //set of things to classify
-    long double* verif_expected,  //set of things to compare against
+    float* verif_samples,     //set of things to classify
+    float* verif_expected,  //set of things to compare against
     size_t num_verif_samples);
 
 /*
@@ -79,19 +79,19 @@ bool sgdNNet(neural_network_t* n_net,
  * Should technically output the gradient, and not just the errors, but
  * that's a lot more to store.
  */
-bool backPropNNet(neural_network_t* n_net, long double* const input,
-    long double* const expected);
+bool backPropNNet(neural_network_t* n_net, float* const input,
+    float* const expected);
 
 /*
  * Verifies the n_net against the verification data
  */
 void verifyNNet(neural_network_t* n_net,
-    long double* const input_data,
-    long double* const expected_data,
+    float* const input_data,
+    float* const expected_data,
     size_t data_size);
 
 //runs net input -> output for classification
-void feedForwardNNet(neural_network_t* n_net, long double* const input);
+void feedForwardNNet(neural_network_t* n_net, float* const input);
 
 bool destroyNNet(neural_network_t* n_net);
 
@@ -99,6 +99,6 @@ bool destroyNNet(neural_network_t* n_net);
 void clearBatchAvg(neural_network_t* n_net);
 
 //from knuth and marsaglia
-long double genRandGauss();
+float genRandGauss();
 
 #endif
