+++ util.h
@@ -4,9 +4,9 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-void printImage(long double* const data, size_t size, size_t width);
+void printImage(float* const data, size_t size, size_t width);
 
-char numToText(long double num);
+char numToText(float num);
 
-size_t getmax(long double* arr, size_t size);
+size_t getmax(float* arr, size_t size);
 #endif
